requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
plotly==5.6.0
streamlit==1.8.1
scikit-learn==0.24.2
pillow==8.3.1